+++ vector_quants/quantizers/radial_quantizer.py
@@ -20,10 +20,10 @@ class RadialQuantizer(BaseVectorQuantizer):
         self.base = base
         num_levels = embed_dim
         levels = [base] * num_levels
-        _levels = torch.tensor(levels, dtype=torch.int32)
+        _levels = torch.tensor(levels, dtype=torch.int64)
         self.register_buffer("_levels", _levels, persistent=False)
         _basis = torch.cumprod(
-            torch.tensor([1] + levels[:-1]), dim=0, dtype=torch.int32
+            torch.tensor([1] + levels[:-1]), dim=0, dtype=torch.int64
         )
         self.register_buffer("_basis", _basis, persistent=False)
 
@@ -63,7 +63,7 @@ class RadialQuantizer(BaseVectorQuantizer):
         # [0, c - 1] -> [0, 1] -> [-1/2, 1/2] -> [-pi/2, pi/2]
         code = torch.sin(torch.pi * (number / d - 0.5))
         # code = number / d
-        indice = (number * self._basis).sum(dim=-1).to(torch.int32)
+        indice = (number * self._basis).sum(dim=-1).to(torch.int64)
 
         return code, indice
